[Review] Removed unused header file
