+++ src/CMakeLists.txt
@@ -274,7 +274,6 @@ list( APPEND SOURCES
       Snap.h
       SoundActivatedRecord.cpp
       SoundActivatedRecord.h
-      SpectralDataDialog.h
       SpectralDataDialog.cpp
       SpectralDataManager.h
       SpectralDataManager.cpp
